libcfnet: bandwidth limit on the network
In a further attempt to prevent CFEngine from clogging my local network,
a throttle is now implemented on the rate we send data out. It is a
global variable, affecting all threads and network interfaces as a sum.

The algo is simple: every time we send data out, we put a penalty on
the time we can use the network again. If we attempt to send earlier than
that, we wait first.

Modified by: Kristian Amlie <kristian.amlie@cfengine.com>