Petty tidy-up in verify_vars.c
Moved a commented-out line, and commentary on it, to after the
declaration of the variable it references.  Split a conditional with
&& on the end of the first line instead of at the start of the second;
and changed x == false to !x.  Made existing_value const by changing
its initialization to a ?: initialization.

Canonicalise indentation in code that reports redefinition of
constants for the AGENT_TYPE_COMMON.  Split some long lines.
Inlined the value of a returned variable where it hasn't been
modified since its constant initializer.