Avoid shadowing a returned variable.
VerifyVarPromise() has a function-scope variable that it returns.  It
also had a pair of nested variables with the same name (result) that
held an int (not a PromiseResult at all) that got returned if
negative.  The latter shadowed the former, so I renamed it; and
returning them doesn't make sense for the return-type, so return the
outer variable.