Make packagesmatching error message into a verbose message.
The inventory policies make sure that the package database is filled,
however the early variable resolution runs before this happens, so
therefore this message appears. Compromise by making it verbose so
you can still see what the problem is if it doesn't work, but it
doesn't pop up as an error.