Add RvalWriteParts() for when we have only a const pointer for item.
When the item we have is a const void * (or other const pointer), we'd
have to cast away its constness in order to put it into an Rval to
pass to RvalWrite(), despite the fact that RvalWrite() only uses it as
a const pointer.  Also, RvalWrite() delegates unpacking of .item to
functions which do nothing else but check the type is right, which we
know it *is* because our calls to them are in relevant cases of a
switch on type; so we really don't need to pack the item into an Rval.