Correction to: Record which promise set each variable.
Our current Promise may be a copy, that'll be free()d at the end of
the current convergence pass; comparing this to the Promise of a later
assignment won't be equal even if it's the original (or another copy
of it); and dereferencing it during reporting shall be regrettable.
Fortunately, a Promise's .org_pp records the original in any copies,
so we can record that instead, bypassing both problems.