Warn if StringWriterClose() return is ignored (hence leaked).
Last leak-fix shows that someone clearly thinks StringWriterClose() is
a good way to close a StringWriter - perhaps not the best choice of
function name, given that this is wrong ! - when actually it's a way
to extract its final value (while closing it).  Therefore callers must
in fact do something with that value - if they ignore the return, they
definitely don't free() it like they should !

I tried doing similar for FileWriterDetach() - but FileWriter(stdout)
and even FileWriter(fd) are commonly used, by callers who have no
business closing stdout or fd; FileWriterDetach() is then the sensible
and proper way to let go after using the Writer.