Fix documentation of parse*array and read*array functions.
The prior documentation was incomplete, confusing and wrong.  The
functions do a rather complex job that it's not easy to document
tersely, but here's my best shot at doing so.

The separate *_ARGS structs for the *stringarrayidx functions were
superfluous - the ones from the plain *stringarray can be used without
change, just as for the *intarray and *realarray functions - and the
one for readstringarrayidx was just plain wrong (talking about an
input string where it should have mentioned a file-name; obvious
copy-and-paste from the readstringarrayidx one).

Fixed a misindentation.