Fix uninitialized memory reads.
In StringWriterWriteLen(), using strlen() and strlcpy() are
problematic when the input comes from read(), which doesn't terminate
its data.  In particular, relying on copying one more byte then the
string length as a way to get a '\0' in place is unsound; there may be
no '\0' at str[len_+1].  So use strnlen(), memcpy() and add the final
'\0' explicitly.

Fixed another place where strnlen() is the right answer, too; and
tidied a comment's layout.