Fix leakage of lists of gid and uid entries.
GetPermissionConstraints() hangs some allocated objects off the .perms
field of an Attributes object.  Attributes objects generally only
carry copies of data elsewhere, but the .owners and .groups members of
.perms now need to be deleted properly when the Attributes object goes
away.  So let's not make GetPermissionConstraints() public any more
(it has no other callers, so let's not solicit more) and we need a
public companion to GetFilesAttributes(), that cleans up its return
when its caller is finished with the Attributes object.  Add that
companion, ClearFilesAttributes(), and call it from the (fortunately
only) one place that calls GetFilesAttributes().  It, in turn, needs
[UG]idListDestroy() functions that tidy away the respective lists.
(How did someone write Rlist2[GU]idList(), returning allocated
pointers, and not notice the need for clean-up functions for those
pointers ?)