Make loop more evidently correct.
Each path round the loop (it continue;s in two places) was setting the
running pointer to the ->next of the previous iteration's item; unify
how that's done so as to make it obviously correct, rather than
obliging the reader to work out what's happening.  Incidentally gives
the optimiser a bit of scope for doing better (because *it's* also
more likely to realise what's actually happening here).