Broke out ExpandList()'s loop body in order to simplify it.
Several paths through the loop body did the same thing as each other,
when computing the expanded value for each entry in the list.  So
break out that expansion into a static function, return early from the
other code paths and let the shared paths all end in the same return.