Renamed NewPackageManager to GetPackageManager.
It's static, so this is a file-local change.  The function doesn't
always create a new package manager; it (hopefully mostly) gets an
existing one from a (global) list and only creates one (that it adds
to the list) if it doesn't find it there already.  Purged some
dangling hspace while I'm here.