Leak-fix: new PromiseIterator()'s ->vars needs a deletion function.
The entries that PromiseIteratorNew() adds to its ->vars sequence are
all CfAssoc objects; some made by NewAssoc(), the rest by a local
piece of code with very similar effect.  In each case, the entry owns
its members, is owned by the sequence and needs to be deleted at some
point.  So tell the sequence to use DeleteAssoc() as its ItemDestroy.