Avoid memory tramplage in PrefixLocalRepository()
There's little point using strlcpy() when first putting data into a
buffer if you're going to follow up with unconditional AddSlash() and
a strcat() into the same buffer; and even less sense in lying to
strlcpy() about the size of the buffer.  Furthermore, if the path thus
assembled was getting truncated in the process, trying to use it was
just going to be misleading.