const-ify PrefixLocalRepository and clean up some callers.
Since it doesn't modify its inputs, using a buffer (idBuff) to hold a
copy of a source string to pass to it was just silly.  When what its
return is being appended to is, by construction, a pointer to the end
of a string, strcpy() is as good as strcat().

Also removed a pointless cast from DeletePackageManagers(), because it
was there (and misleading - and distracting).