Fail PrependPackageItem() if its list == NULL.
If this PackageItem ** is NULL, there's nowhere to store the
PackageItem* we allocate, so this allocation is a leak.  While we're
here, to determine whether a string is empty you only need to look at
its first byte; no need to run past that looking for an index at which
you might find a '\0' in order to compare this index to 0.