Assorted clean-up in rlist.c
Mostly efforts to make it easier to read the code.  Added a few
assertions.  Made some locals more local.  Turned some for loops with
empty bodies into while loops where (IMO) it's easier to see what the
point of the loop is that way.  Split some long lines.  Use parameter
in place of a local variable copied from it in a few cases.