Change csv_parser max size to 4kb
csv parser is based on static buffers, and I want it to be aligned
with CF_BUFSIZE but didn't want to introduce cf3.defs.h as dependency,
so I used a number.

This should be based on malloc and not on static buffer but don't have
time at the moment to make bigger change.