Eliminated some clumsiness in instrumentation code.
When a flag is initialized true, set false by some failure-handling
code and then conditioned on, it can be eliminated in favour of just
having an else to the failure-handler.  Arithmetic with doubles really
doesn't need to cast anything like so much to doubles; anything
multiplied by an overt double (e.g. 1e3) is implicitly converted to
double anyway.