Fix testcase checking for zero-size files
In the msys env on win, /dev/null doesn't exist, and since we're not
interested in it per se, but just want to have something that is known
to be zero-sized, amend the test infrastructure accordingly:

	1. add zero-byte dummy_etc/null
	2. add a stanza to common bundle G default.cf.sub to deal with
	it
	3. amend 071.cf to replace /dev/null with a variable

That fixes half of 071.cf. It fails on win in yet another way, sinze
filesize() will return a different integer on nix than on win, due to
CRLF characters. Add crude os classes to check bundle. I was hoping this
could have been solved in a more elegant way, but was informed there
isn't any.