Finish the testcase section dealing with Windows
Feed a Windows(R) Registry(TM) path to registryvalue(), one such that
always exists on any Win install, regardless of customizations, namely
ProgramFiles.

Non-win-native humans reading this tc in the future should probably take
note that valueid in registryvalue(damer,valueid) is the *name* of the
key, while the value is what this function actually returns.