Make SV.logconns actually mean something.
It controls logging of reporting of connections to the server: if it's
set, we log them - otherwise, we log them - in each case at the info
level.  I decided that we could at least have it control the logging
level used: when it's not set I've downgraded the logging level to
verbose.