packagesmatching() needs work — see issue #4741
The data container returned by packagesmatching() is now populated with
an executable name with an extension (printf.bat), whereas one without
is expected.

Quite apart from that, even if the actual and expected containers have
the same content, line ending differences will ruin the comparison, so
the tc will fail anyway.

Suppress testing of this tc on win.