Fix up some #if-ery to avoid unreachable code.
Some avahi-specific code ended in an exit() and was followed by more
code, which would thus be unreachable when the relevant avahi defines
are enabled.  The code involved also repeated a pair of #if tests in
three places; so make the first set a local define the other two can
use as a simpler test; have a #else clause for the formerly unreached
code; and share the common exit() between the two branches.