Methods that do nothing are also verified/kept
Fixes Redmine #4852. If a bundle invoked via a methods promise
results only in skipped promises, then the methods promise is kept.

That this issue exists with the test case in the redmine ticket
suggests that file promises return inconsistent results here. A
promise to create a file when that file exists should return NOOP,
not SKIPPED.