Fix leaky potential list-mangling bug.
DeleteItemGeneral's item deletion code depended on a variable last
that was meant to record the node with last->next == ip; however, due
to the use of continue (eliminated last commit, but faithfully
preserving what the code does) to skip the loop body if ip->name is
NULL, the code to update last would be skipped.  Consequently last
could be NULL when ip wasn't the list's head; and, when non-NULL,
might have several steps of ->next, via entries with NULL name, before
ip.  Changing its ->next to ip->next would drop the intermediate
entries, and leave them pointing at ip after it's free()d, with
potentially scary results.  Skipping the update of a relevant item's
->next when last was NULL and ip wasn't the list head would leave ip
in the list after it's deletion.

So move setting of last = ip out of the conditional (previously
implicit in the continue).  This ensures we can simply assert(last)
instead of checking it, in the place where it's needed; the same
code's check on ip was superfluous anyway, so assert(ip), too; and
avoid a bunch of code duplication around free()ing ip and its members.