Add ItemIsInList() for use in an assertion.
Checks for an Item being in a list (in contrast to IsItemIn, which
checks for the list having an entry whose ->name strcmp()s the same as
a given string); this is needed in ConcatLists() to check we're not
creating a cycle.  It could also be used in reverse and delete
functions to check for cycles there, but at quadratic cost, where
concat only incurs linear cost (which it already had anyway).