Fixes and clean-up in ItemList2CSV_bound().
It's meant to concatenate a bunch of strings, joining them with a
separator.  However, it actually copied each string in turn to the
start of the buffer, while writing separators, and appending a final
'\0', in all the right places, had the content but been between them.

It also checked whether the space remaining was more than two less
than the length of the string it was about to copy into that space;
its comments make clear that it meant to leave two extra spaces, but
it actually left two too few.  It doesn't actually need two anyway, so
would have truncated the last byte if the buffer were just exactly big
enough.  Check separately, when appending a separator, for whether
there's space for it.

Its documentation also wrongly claimed that the string stored in the
buffer would have length equal to the buffer size on truncation (the
actual strlen() shall be one less) when it meant that the *return* is
the buffer size - hence *bigger* than strlen(buf), since we really do
'\0'-terminate - when we truncate.

... and re-write to use a pointer to where to write next, while
keeping track of remaining space as we go, with the -1 for '\0' taken
into account at the start.  It's just simpler that way ...