Simplify SplitString.
Not every loop wants to be a for; if you find yourself having to undo
the increment step in some paths through the loop body, it's a good
hint that it doesn't.  Restructured the loop body as a simple
if/else-if/else chain in which a separator at the start is handled
just as a later unescaped separator, instead of duplicating that code.