Simplify SplitStringAsItemList().
Instead of using snprintf() to insert a character in a string, set the
string up sensibly in the first place so that we can simply copy the
character into the right place in it.  Instead of memset()ing 4k
in hopes that a failing sscanf won't touch the destination buffer,
actually check sscanf's return.