Potential nightmare-fix in DeleteItem().
When searching a list for the entry whose ->next is the one we're
about to delete, we'd stop on the last item, which doesn't have the
given entry as ->next (it has NULL).  We would then assume that, as it
(as opposed to its ->next) is non-NULL, we should hang our item's
->next off it, where it never did belong.  If the item we're deleting
had our supposed list head somewhere down->next of it, this could have
created a cycle.  So let the loop variable fall off the end of the
loop as NULL and do at least assert that a non-NULL alleged
predecessor has the node to be deleted as its ->next, before fixing
that.