Eliminate ByteSizeList in favour of ItemListSize.
Both computed the sum of lengths of strings in a list.  The sole
difference is that the latter checks for, and skips, NULL entries, on
which the former would crash.  Code outside core uses the latter, only
code inside core uses the former, so this change is self-contained to
core (easier to git administer).

Also make ItemList2CSV() robust against NULL entries, now that it's
delegating size computation to ItemListSize().