Fix agents downloading policy every 5 min.
Multiple causes led to this. First, the addition of a checksum in
cf_promises_validated. Since we don't guarantee that the masterfiles
directory on the hub and the inputs directory on the agent are the
same, the checksums could easily be different. This would lead to the
agent regenerating the cf_promises_validated, believing it to be
wrong, and then overwriting it again with the file from the server, in
an endless loop. Fixed by relying on timestamp once again, not
checksum.

Also, move the generated file back to masterfiles. Because a
cf_promises_validated is downloaded from the server, its modified
timestamp will be newer than the timestamp contained within it, which
will trigger regeneration. Next time around the same thing happens.

Also moved the internal validated_at config value out of the
GenericAgentIsPolicyReloadNeeded() function to avoid a const cast.