Simplify GetAcpi().
Make various variables more local.
Make a conditional more readable.

Eliminate a buffer and the sprintf into it where all we needed was to
search for one digit in a string; adding the (singled-digit) int to
'0' will get the digit, with strchr replacing strstr to find it.
Replace a repetitive switch where indexing into a simple array can do
the same job more straightforwardly.

Discovered that a Log() was trying to output the contents of a path
buffer that hadn't actually been initialized; it should have contained
the name of the directory we'd just failed to open, so initialized it
that way.