warning-fix: Correct signature of SaveXmlCallback().
It's used as a SaveCallbackFn, so needs to accept a NewLineMode as
third parameter, even if it does ignore it.  It should also be a
static function (nothing outside its file uses it); and we may as well
simplify the #if-ery around it and its one caller.

(cherry picked from commit ddd0cdc4fd431c54f97ae35467ba604f09a0dfa2)