Simplification and clarification in SearchAndReplace().
Simplify the loop logic from a for (;;) with a conditional return at
the start of its body to a suitable while (albeit with assignment in
its condition), so that we can have return at the end of the body.
Declare variables after the early returns that don't use them.
Only compute length of search string once.
Rename variables.  Simplify arithmetic.