Identification dialog now only needs one roundtrip
Changed identification dialog to send two lines at once from the client
(protocol version and client identity) so that we save one network
roundtrip.

Had to bring back TLSRecvLines() for that purpose (why was it removed in
the first place?), to make sure that no bytes will be lost when
receiving, even if TLS squeezes multiple lines in one record.