Refactor: consolidate the PlatformContext-related externs.
Move their declarations out of cf3.*.h, to a header file of their own,
and move an associated definition from cf3globals.c to classes.c so
that they're all in one place.  Renamed classes.c to systype.c and
named the header systype.h, since this name more faithfully relates to
the actual function of PlatformContext (it says what type of system
we're running on).  Adjust includes and make-files to match.