redmine3974: add "Sun solaris" platform type.
This distinguishes solaris with BSD-style tools in /usr/ucb/ from more
modern versions where the system tools are BSD-compatible.  For now,
there is no difference; this is preparation for adding differences.
Documented what various platforms are a bit more clearly.

Note on the name: although Oracle took over before SunOS 5.10 was out,
the versions with /usr/ucb/ were part of the Sun era; and we need a
simple variant on the name to distinguish the two cases.  Using prefix
OLD has the perennial problem that, some day, what's currently new
shall be old.  It seems unlikely that Sun shall ever be an independent
entity again, so using its name looks like a better choice.