Mark sprintf as deprecated.
A warning will be now output on every usage of that function on
GNU/linux. sprintf() is the only function for which we do *not* provide
a C99 libcompat replacement. While it's trivial to do it, I chose to
mark the function as *deprecated* because there is no good reason to use
it in place of snprintf().

In addition there is one good reason to *not* provide libcompat wrapper
for sprintf() and leave it as it is: so that we can call the specific
function to force invoking the underlying libc implementation. This is
the way the snprintf() test works (see libcompat/snprintf.c), it
compares its own results with libc's sprintf() results.