file_writer_test was refactored.
The reason is that previously we were redefining libc's fwrite() to
change the output location of WriterWrite(). However, on Solaris 9
fwrite() is also used in our printf() replacement which is called for
other reasons in the tests, so random crashes/behaviour was happening.

The test is now including the writer.c source file directly, which
overrides fwrite() only if #ifdef CFENGINE_TEST.