redmine5411: conn == NULL is used to indicate local "connection".
The callers of cf_stat() use it with this meaning; but it was only
testing fc.servers, which can be set to something other than
"localhost" (e.g. a list with "localhost" as the entry that's actually
being used, after failure to contact earlier entries) when conn is
NULL.  Various other functions made the same over-simplification.

Added a simple acceptance test for the resulting crasher.

Whenever the old test triggered, conn is in fact NULL (I wrote some
assertions to check that, at least in an acceptance test run); so I
simplified these conditions to just check conn (and assert the residue
in the else clause).