Fix package_list_update_ifelapsed having no effect on "apt-get update"
Redmine #3981, #5339.

It is documented as having the described effect, but does not.

The timestamp for the last update is now maintained by a timestamp
file, /var/cfengine/state/software_update_timestamp_<manager>, where
<manager> is the package executable that is being used for the current
promise.
Ideally it should be integrated with the database, but we want the low
risk solution right now, and we plan to replace this code anyway.