Clean up in MakeParentDirectory().
Eliminate a Bad Cast by using a const char * local to traverse the
const char * parameter; and use a different (non-const) local for the
one other place that local was used before.

Using strncpy to copy part of a string leaves it unterminated; use
memcpy (to save checks for its end) and terminate it correctly.

Make variable declarations more local, split long lines, unfold a few
else { if ... } to else if (thereby losing some indentation; view with
git show -b or add a ?w=1 to the github URL for best results).